update missing components
