+++ Web_Frontend/genfit_frontend/src/pages/goal/goal_page.css
@@ -126,7 +126,6 @@
 .goal-card:hover {
     border-color: #d1d5db;
     box-shadow: 0 10px 25px -3px rgba(0, 0, 0, 0.1);
-    transform: translateY(-2px);
 }
 
 .goal-card-header {
@@ -293,6 +292,67 @@
     grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
     gap: 1.5rem;
     margin-bottom: 2rem;
+}
+
+/* Goal Tabs */
+.goal-tabs {
+    display: flex;
+    gap: 0.5rem;
+    margin-bottom: 1.5rem;
+    border-bottom: 2px solid #f3f4f6;
+    overflow-x: visible;
+    border-bottom: 4px solid #800000;
+}
+
+.goal-tab {
+    background: none;
+    border: none;
+    padding: 0.75rem 1rem;
+    border-radius: 8px 8px 0 0;
+    font-size: 0.875rem;
+    font-weight: 500;
+    color: #6b7280;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    white-space: nowrap;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    position: relative;
+}
+
+.goal-tab:hover {
+    background: #f9fafb;
+    color: #374151;
+}
+
+.goal-tab.active {
+    background: #800000;
+    color: white;
+    font-weight: 600;
+}
+
+.goal-tab.active::after {
+    content: '';
+    position: absolute;
+    bottom: -2px;
+    left: 0;
+    right: 0;
+    height: 2px;
+    background: #800000;
+}
+
+.goal-count {
+    font-size: 0.75rem;
+    background: rgba(255, 255, 255, 0.2);
+    padding: 0.125rem 0.375rem;
+    border-radius: 10px;
+    font-weight: 500;
+}
+
+.goal-tab:not(.active) .goal-count {
+    background: #e5e7eb;
+    color: #6b7280;
 }
 
 .stat-card {
