+++ Web_Frontend/genfit_frontend/src/pages/home/home_page.css
@@ -193,7 +193,6 @@
 }
 
 .goal-card:hover {
-  transform: translateY(-2px);
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
